LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/radio-focus-navigation-disabled-expected.txt

Outside form
disabled checked  enabled 1  enabled 2
Outside form

FAIL Should be able to tab focus in radio group even when the checked radio button is disabled. assert_equals: expected Element node <input type="radio" name="radio" id="b"></input> but got Element node <body><form>
 <label><input type="radio" name="start" id=...
